server_backend/requirements.txt
flask
torch
transformers
Pillow
ultralytics
opencv-python
